+++ frontend/src/modules/attachments/attachments-table/table.tsx
@@ -5,7 +5,6 @@ import { Paperclip } from 'lucide-react';
 import type { RowsChangeData } from 'react-data-grid';
 import { useTranslation } from 'react-i18next';
 import { useBreakpoints } from '~/hooks/use-breakpoints';
-import { useDataFromSuspenseInfiniteQuery } from '~/hooks/use-data-from-query';
 import type { AttachmentSearch, AttachmentsTableProps } from '~/modules/attachments/attachments-table';
 import { useColumns } from '~/modules/attachments/attachments-table/columns';
 import { attachmentsQueryOptions } from '~/modules/attachments/attachments-table/helpers/query-options';
@@ -15,6 +14,7 @@ import ContentPlaceholder from '~/modules/common/content-placeholder';
 import { DataTable } from '~/modules/common/data-table';
 import type { ColumnOrColumnGroup } from '~/modules/common/data-table/columns-view';
 import { dialog } from '~/modules/common/dialoger/state';
+import { useDataFromSuspenseInfiniteQuery } from '~/query/hooks/use-data-from-query';
 import { useAttachmentUpdateMutation } from '~/query/mutations/attachments';
 import { useUserStore } from '~/store/user';
 import type { Attachment, BaseTableMethods, BaseTableProps } from '~/types/common';
