+++ frontend/src/modules/system/requests-table/table.tsx
@@ -1,11 +1,11 @@
 import { Bird } from 'lucide-react';
 import { forwardRef, memo, useEffect, useImperativeHandle } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useDataFromSuspenseInfiniteQuery } from '~/hooks/use-data-from-query';
 import ContentPlaceholder from '~/modules/common/content-placeholder';
 import { DataTable } from '~/modules/common/data-table';
 import type { RequestsSearch } from '~/modules/system/requests-table';
 import { requestsQueryOptions } from '~/modules/system/requests-table/helpers/query-options';
+import { useDataFromSuspenseInfiniteQuery } from '~/query/hooks/use-data-from-query';
 import type { BaseTableMethods, BaseTableProps, Request } from '~/types/common';
 
 type BaseRequestsTableProps = BaseTableProps<Request, RequestsSearch>;
