+++ frontend/src/modules/users/users-table/index.tsx
@@ -3,7 +3,6 @@ import { config } from 'config';
 import { Suspense, lazy, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import type { z } from 'zod';
-import { useMutateQueryData } from '~/hooks/use-mutate-query-data';
 import useSearchParams from '~/hooks/use-search-params';
 import { useUserSheet } from '~/hooks/use-user-sheet';
 import { showToast } from '~/lib/toasts';
@@ -13,6 +12,7 @@ import DeleteUsers from '~/modules/users/delete-users';
 import InviteUsers from '~/modules/users/invite-users';
 import { useColumns } from '~/modules/users/users-table/columns';
 import { UsersTableHeader } from '~/modules/users/users-table/table-header';
+import { useMutateQueryData } from '~/query/hooks/use-mutate-query-data';
 import { usersKeys } from '~/query/query-key-factories';
 import { UsersTableRoute, type usersSearchSchema } from '~/routes/system';
 import type { BaseTableMethods, User } from '~/types/common';
