+++ frontend/src/modules/users/users-table/table.tsx
@@ -4,13 +4,13 @@ import { updateUser } from '~/api/users';
 
 import type { RowsChangeData } from 'react-data-grid';
 import { useTranslation } from 'react-i18next';
-import { useDataFromSuspenseInfiniteQuery } from '~/hooks/use-data-from-query';
-import { useMutateQueryData } from '~/hooks/use-mutate-query-data';
 import { useMutation } from '~/hooks/use-mutations';
 import { showToast } from '~/lib/toasts';
 import { DataTable } from '~/modules/common/data-table';
 import type { UsersSearch } from '~/modules/users/users-table';
 import { usersQueryOptions } from '~/modules/users/users-table/helpers/query-options';
+import { useDataFromSuspenseInfiniteQuery } from '~/query/hooks/use-data-from-query';
+import { useMutateQueryData } from '~/query/hooks/use-mutate-query-data';
 import { usersKeys } from '~/query/query-key-factories';
 import type { BaseTableMethods, BaseTableProps, User } from '~/types/common';
