+++ frontend/src/query/hooks/use-map-query-data-to-rows.tsx
