+++ frontend/src/query/mutations/members.ts
@@ -5,10 +5,10 @@ import { config } from 'config';
 import { t } from 'i18next';
 import { queryClient } from '~/lib/router';
 import { showToast } from '~/lib/toasts';
+import { formatUpdatedData, getCancelingRefetchQueries, getQueries, getQueryItems, handleNoOldData } from '~/query/helpers/mutate-query';
 import { membersKeys } from '~/query/query-key-factories';
 import type { ContextProp, InfiniteQueryData, QueryData } from '~/query/types';
 import type { Member, Membership } from '~/types/common';
-import { formatUpdatedData, getCancelingRefetchQueries, getQueries, getQueryItems, handleNoOldData } from '~/utils/mutate-query';
 
 type MemberQueryData = QueryData<Member>;
 type InfiniteMemberQueryData = InfiniteQueryData<Member>;
